Worked on the sidebar -> it kinda works
The 'hover your mouse' over the links kinda works, but has some
odd behaviour when you click on it and it does change to black.

I've also been working on the top navbar, and it does as intended.
The items in the top navbar overflow to the bottom (instead of
collapsing to a menu) when the size becomes small.